Munny Pouch overflow bugfix

+++ GoA RAM Non-Rando.lua
@@ -722,12 +722,12 @@ if true then --No Valor, Wisdom, Master, or Final
 end
 --Munny Pouch (Olette)
 while ReadByte(Save+0x363C) > ReadByte(Save+0x35C4) do
-	WriteShort(Save+0x2440,ReadShort(Save+0x2440)+5000)
+	WriteInt(Save+0x2440,ReadInt(Save+0x2440)+5000)
 	WriteByte(Save+0x35C4,ReadByte(Save+0x35C4)+1)
 end
 --Munny Pouch (Mickey)
 while ReadByte(Save+0x3695) > ReadByte(Save+0x35C5) do
-	WriteShort(Save+0x2440,ReadShort(Save+0x2440)+5000)
+	WriteInt(Save+0x2440,ReadInt(Save+0x2440)+5000)
 	WriteByte(Save+0x35C5,ReadByte(Save+0x35C5)+1)
 end
 --Alternate Party Models (adding new UCM using MEMT causes problems when shopping)
